requirements.txt
# direct requirements
# Web框架
fastapi==0.121.2
uvicorn[standard]==0.24.0
starlette==0.49.3

# 前端框架
streamlit==1.52.1

# AI框架
langchain==1.0.7
langchain-core==1.2.1
langchain-community  # 包含 ChatTongyi, ChatOpenAI, ChatOllama 等
langchain-openai==1.0.3
langchain-google-genai==3.1.0

# 模型支持
openai==2.8.1
ollama-python==0.1.2
dashscope

# 工具和工具包
httpx==0.26.0
requests==2.32.5

# 配置管理
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.0.0

# 数据处理
pandas==2.3.3
numpy==1.26.4

# 其他依赖
anyio==4.11.0
click==8.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
tqdm==4.67.1

pytest
protobuf
pydantic-settings
supabase